Add Dockerfile and modify build script
Introduced a new Dockerfile and modified the existing build.sh script. The Dockerfile sets up the environment for the ROS package and the build script removes the `--no-cache` option to optimize the Docker build process. Renamed the lidar_camera_fusion package to lidar-camera-fusion for consistency and made updates to parameters and topic subscriptions to improve node performance.