Refactor configuration and add multi-launch file
This commit includes a resctructuring of the configuration files. It also renames namespaces in `vlp16OnImg_multi.launch` for better clarity. Moreover, unnecessary code in `lidar_camera_node.cpp` has been removed and there are changes in parameters, topic names and small adjustments accompanying the introduction of the multi-launch capability.